+++ data-jpa/src/main/java/net/optionfactory/spring/data/jpa/filtering/WhitelistSortingSpecificationAdapter.java
@@ -7,17 +7,17 @@ import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import java.util.Map;
 import java.util.stream.Stream;
+import net.optionfactory.spring.data.jpa.filtering.filters.spi.Filters.Traversal;
 import net.optionfactory.spring.data.jpa.filtering.filters.spi.Sorters;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
-import org.springframework.data.jpa.repository.query.QueryUtils;
 
 public class WhitelistSortingSpecificationAdapter<T> implements Specification<T> {
 
     private final Sort requested;
-    private final Map<String, String> allowed;
+    private final Map<String, Traversal> allowed;
 
-    public WhitelistSortingSpecificationAdapter(Sort requested, Map<String, String> allowed) {
+    public WhitelistSortingSpecificationAdapter(Sort requested, Map<String, Traversal> allowed) {
         this.requested = requested;
         this.allowed = allowed;
     }
@@ -26,7 +26,7 @@ public class WhitelistSortingSpecificationAdapter<T> implements Specification<T>
     public Predicate toPredicate(Root<T> root, CriteriaQuery<?> query, CriteriaBuilder criteriaBuilder) {
         query.orderBy(Stream.concat(
                 query.getOrderList().stream(),
-                QueryUtils.toOrders(Sorters.validateAndTransform(root.getJavaType(), requested, allowed), root, criteriaBuilder).stream()
+                Sorters.orders(root, criteriaBuilder, requested, allowed).stream()
         ).toArray(i -> new Order[i]));
         return null;
     }