+++ data-jpa/src/main/java/net/optionfactory/spring/data/jpa/filtering/filters/spi/InvalidSortConfiguration.java
@@ -1,0 +1,10 @@
+package net.optionfactory.spring.data.jpa.filtering.filters.spi;
+
+import jakarta.persistence.metamodel.EntityType;
+
+public class InvalidSortConfiguration extends IllegalStateException {
+
+    public InvalidSortConfiguration(String sorterName, EntityType<?> entity, String reason) {
+        super(String.format("in sorter %s@%s: %s", sorterName, entity.getJavaType().getSimpleName(), reason));
+    }
+}